+++ src/index.ts
@@ -2,5 +2,6 @@ export interface RFIFactory<T> {
   make: (...args: any[]) => T;
 }
 
-class ServiceFactoryMap extends Map<string, RFIFactory<any>> {}
-export const ServiceFactories = new ServiceFactoryMap();
+export abstract class ServiceFactoryMap extends Map<string, RFIFactory<any>> {}
+class SingletonMap extends ServiceFactoryMap {}
+export const ServiceFactories = new SingletonMap();